✨ credentials auth
