+++ src/app/(main)/auth/signin/page.tsx
@@ -1,4 +1,4 @@
-import { LoginDialog } from '@/auth/LoginDialog'
+import { LoginForm } from '@/auth/LoginForm'
 import { getIsLoggedIn } from '@/auth/getMyUser'
 import { Card, CardContent } from '@/components/ui/card'
 import { redirect } from 'next/navigation'
@@ -9,16 +9,15 @@ export default async function Page({
   searchParams: { redirect?: string }
 }) {
   const isLoggedIn = await getIsLoggedIn()
-  console.log('isLoggedIn', isLoggedIn)
   if (isLoggedIn) {
-    console.log('redirectPath', redirectPath)
     redirect(redirectPath ? decodeURIComponent(redirectPath) : '/')
   }
+
   return (
     <>
       <Card className="self-center w-full max-w-md flex flex-col gap-4">
         <CardContent className="flex flex-col gap-4 pt-6">
-          <LoginDialog />
+          <LoginForm />
         </CardContent>
       </Card>
     </>
